backend/requirements.txt
# ==========================================
# FastAPI & Web Framework
# ==========================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ==========================================
# Database
# ==========================================
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ==========================================
# Redis
# ==========================================
redis==5.0.1

# ==========================================
# Data Validation
# ==========================================
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ==========================================
# LLM APIs
# ==========================================
anthropic==0.8.1
openai==1.7.2
httpx==0.26.0

# ==========================================
# Authentication & Security
# ==========================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ==========================================
# CORS & WebSocket
# ==========================================
python-socketio==5.11.0
websockets==12.0

# ==========================================
# Utils
# ==========================================
python-dateutil==2.8.2
pytz==2023.3

# ==========================================
# Audio Processing
# ==========================================
pydub==0.25.1

# ==========================================
# Testing
# ==========================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# ==========================================
# Logging & Monitoring
# ==========================================
structlog==24.1.0

# ==========================================
# RAG Dependencies
# ==========================================
chromadb==0.4.22
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2

# ==========================================
# Vocode (будет добавлено в фазе 5)
# ==========================================
# vocode==0.1.0
